refactor(frontend): move request types to request entity_by dev-WAKR93-#294
