+++ frontend/src/entities/requests/api/requests.types.ts
@@ -1,0 +1,10 @@
+import { RequestStatus } from "@/entities/auth/model/types";
+
+export interface CreateRequestDto {
+  sender: string;
+  receiver: string;
+  offeredSkill: string;
+  requestedSkill: string;
+  status?: RequestStatus;
+  isRead?: boolean;
+}
